chore: move equity shares to common
