+++ core/execution/common/equity_shares_snapshot.go
@@ -10,7 +10,7 @@
 // of this software will be governed by version 3 or later of the GNU General
 // Public License.
 
-package future
+package common
 
 import (
 	"sort"
