+++ protos/sources/vega/snapshot/v1/snapshot.proto
@@ -407,26 +407,6 @@ message EquityShare {
   string p_mvp = 5;
 }
 
-message SpotEquityShareLP {
-  string id = 1;
-  string stake = 2;
-  string buy_stake = 3;
-  string sell_stake = 4;
-  string share = 5;
-  string avg = 6;
-  string vshare = 7;
-  string buy_vshare = 8;
-  string sell_vshare = 9;
-}
-
-message SpotEquityShare {
-  string mvp = 1;
-  bool opening_auction_ended = 2;
-  repeated SpotEquityShareLP lps = 3;
-  string r = 4;
-  string p_mvp = 5;
-}
-
 message FeeSplitter {
   int64 time_window_start = 1;
   string trade_value = 2;
@@ -446,7 +426,7 @@ message SpotMarket {
   string last_mid_ask = 9;
   string last_market_value_proxy = 10;
   int64 last_equity_share_distributed = 11;
-  SpotEquityShare equity_share = 12;
+  EquityShare equity_share = 12;
   string current_mark_price = 13;
   FeeSplitter fee_splitter = 14;
   int64 next_mark_to_market = 15;
